fix and remove tests that used the selectCurrentUser selector
